Move generally useful helpers to varutils
